refactor: modularize hotkey handling
